requirement.txt
numpy
pandas
matplotlib
scikit-learn
opencv-python
tensorflow
keras
Pillow